fix metatags not showing up
