+++ pages/dex.tsx
@@ -1,29 +1,52 @@
 import type { NextPage } from 'next';
-
+import Head from 'next/head';
 import {
   getStoryblokApi,
   StoryblokComponent,
+  StoryData,
   useStoryblokState,
 } from '@storyblok/react';
-import Head from 'next/head';
-import MetaTags, { IMetaTags } from '../src/components/Atoms/MetaTags';
 
-// @ts-ignore
-// eslint-disable-next-line react/prop-types
-const Dex: NextPage = ({ story }) => {
+import { DEFAULT_HYDRANET_COVER_URL } from '../src/constants';
+import { IMetaTags } from '../src/storyblok/models/IMetaTags';
+
+const Dex: NextPage<{ story: StoryData }> = ({ story }) => {
   story = useStoryblokState(story);
   // @ts-ignore
-  // eslint-disable-next-line react/prop-types
   const { MetaTagsBlok }: Array<IMetaTags> = story.content;
+  const { title, description, type, url, image, locale } = MetaTagsBlok[0];
 
   return (
     <>
       <Head>
-        {/* eslint-disable-next-line react/prop-types */}
         <title>{MetaTagsBlok[0]?.title || 'Hydranet'}</title>
-        <MetaTags {...MetaTagsBlok[0]} />
+
+        <title>{MetaTagsBlok[0]?.title || 'Hydranet'}</title>
+        <meta name="title" content={title} />
+        <meta name="description" content={description} />
+
+        <meta property="og:type" content={type} />
+        <meta property="og:url" content={url} />
+        <meta property="og:title" content={title} />
+        <meta property="og:description" content={description} />
+        <meta
+          property="og:image"
+          content={image?.filename || DEFAULT_HYDRANET_COVER_URL}
+        />
+        <meta property="og:image:alt" content={image?.alt} />
+        <meta property="og:locale" content={locale} />
+
+        <meta property="og:site_name" content="Hydranet" />
+
+        <meta property="twitter:card" content="summary" />
+        <meta property="twitter:url" content={url} />
+        <meta property="twitter:title" content={title} />
+        <meta property="twitter:description" content={description} />
+        <meta
+          property="twitter:image"
+          content={image?.filename || DEFAULT_HYDRANET_COVER_URL}
+        />
       </Head>
-      {/* eslint-disable-next-line react/prop-types */}
       <StoryblokComponent blok={story.content} />;
     </>
   );
