+++ src/constants/index.ts
@@ -1,39 +1,5 @@
-export const UNISWAP_BUY_LINK =
-  'https://app.uniswap.org/#/swap?inputCurrency=ETH&outputCurrency=0xF4fe727C855c2D395852ca43F645caB4b504Af23&chain=arbitrum';
-export const SUSHISWAP_BUY_LINK =
-  'https://app.sushi.com/swap?inputCurrency=0xF4fe727C855c2D395852ca43F645caB4b504Af23&outputCurrency=ETH';
-export const WHITEBIT_REF_LINK =
-  'https://whitebit.com/referral/0b1df51a-c47d-4665-b126-5ba687a9756a';
-
-export const BOND_BUY_LINK = 'https://app.hydranet.ai/#/bonds';
-
-export const HYDRANET_APP = 'https://app.hydranet.ai/#/dashboard';
-
 export const HYDRANET_MEDIUM_FETCH_URL =
   'https://aiutsa6z6b.execute-api.us-east-1.amazonaws.com/default/hydranet-website-fetch-medium';
 
-type AvailableBuilds = 'mainnet' | 'testnet';
-export type DexDownloadLink = {
-  [K in AvailableBuilds]: {
-    windows: string;
-    linux: string;
-    macos?: string;
-  };
-};
-export const DEX_DOWNLOAD_LINKS: DexDownloadLink = {
-  mainnet: {
-    windows:
-      'https://hydranet.s3.eu-central-1.amazonaws.com/light-wallet/staging/win/installer/Hydranet.zip',
-    linux:
-      'https://hydranet.s3.eu-central-1.amazonaws.com/light-wallet/staging/linux/installer/Hydranet.zip',
-  },
-  testnet: {
-    windows:
-      'https://hydranet.s3.eu-central-1.amazonaws.com/light-wallet/staging/win/installer/Hydranet.zip',
-    linux:
-      'https://hydranet.s3.eu-central-1.amazonaws.com/light-wallet/staging/linux/installer/Hydranet.zip',
-  },
-};
-
-export const SSUI_YOUTUBE_BTC_ETH = 'https://youtu.be/8Tu0B-Ikz4g';
-export const SSUI_YOUTUBE_ETH_BTC = 'https://youtu.be/8n-XegXsmxw';
+export const DEFAULT_HYDRANET_COVER_URL =
+  'https://a.storyblok.com/f/154370/700x445/3f1b4e8671/hydra-meta.jpg';
